+++ PA4/kernel.cl
@@ -1,5 +1,5 @@
 __kernel void matrixMultiply(
-    __global const float *A, __global const float *B, __global float *C,
+    __global const int *A, __global const int *B, __global int *C,
     const unsigned int numARows, const unsigned int numAColumns,
     const unsigned int numBRows, const unsigned int numBColumns,
     const unsigned int numCRows, const unsigned int numCColumns) {
